feat : utils 폴더에 formatDate 추가 photoCardInfo 타입수정 photoCardList Props 에 데이터 입력받는 형식으로 변경
