+++ src/constants/photoCardType.ts
@@ -1,0 +1,10 @@
+export type PhotoCardInfoType = {
+  id: number;
+  ImageUrl: string;
+  Area: string;
+  Like: number;
+  Text: string;
+  Writer: string;
+  Date: Date;
+  Tags: string[];
+};
