+++ src/containers/InformationPage/InformationPage.tsx
@@ -3,23 +3,24 @@ import React from 'react';
 import { Category } from './components';
 import { FixedButton, Notice, PhotoCardList, Search } from 'src/components';
 import styled from 'styled-components';
+import { photoCardInfo } from 'src/dummy/photoCardInfo';
 export const InformationPage: NextPage = () => {
   return (
     <Container>
       <Search />
       <Notice />
       <FixedButton />
-      <Category category="여행" pathName='/travel'>
-        <PhotoCardList type="MainPage" />
+      <Category category="여행" pathName="/travel">
+        <PhotoCardList type="MainPage" photoCardInfo={photoCardInfo} />
       </Category>
-      <Category category="음식점" pathName='/shop'>
-        <PhotoCardList type="MainPage" />
+      <Category category="음식점" pathName="/shop">
+        <PhotoCardList type="MainPage" photoCardInfo={photoCardInfo}/>
       </Category>
-      <Category category="숙소" pathName='/hotel'>
-        <PhotoCardList type="MainPage" />
+      <Category category="숙소" pathName="/hotel">
+        <PhotoCardList type="MainPage" photoCardInfo={photoCardInfo}/>
       </Category>
-      <Category category="기타" pathName='/other'>
-        <PhotoCardList type="MainPage" />
+      <Category category="기타" pathName="/other">
+        <PhotoCardList type="MainPage" photoCardInfo={photoCardInfo}/>
       </Category>
     </Container>
   );
