+++ src/dummy/photoCardInfo.ts
@@ -1,20 +1,15 @@
-export type PhotoCardInfoType = {
-  ImageUrl: string;
-  Area: string;
-  Like: number;
-  Text: string;
-  Writer: string;
-  Date: string;
-  Tags: string[];
-};
+import { PhotoCardInfoType } from "src/constants/photoCardType";
+let id = 1;
+let Like = 1;
 export const photoCardInfo: PhotoCardInfoType[] = [
   {
+    id: id++,
     ImageUrl: 'favicon.ico',
     Area: '인천 구월동',
-    Like: 12,
+    Like: Like++,
     Text: '이건 뭔사진인지 당최 모르겠네 가나다라마바사아자차카타파하 아직도 남았는데 가나다라마바사 두줄이상이면 점으로 돼야하는데 이게 될라나',
     Writer: '글쓴이닉네임',
-    Date: '2022-04-08',
+    Date: new Date(),
     Tags: [
       '#인천',
       '#여행',
@@ -22,12 +17,13 @@ export const photoCardInfo: PhotoCardInfoType[] = [
     ],
   },
   {
+    id: id++,
     ImageUrl: 'favicon.ico',
     Area: '인천 구월동',
     Like: 12,
     Text: '한줄이면 이렇게 출력',
     Writer: '글쓴이닉네임',
-    Date: '2022-04-08',
+    Date: new Date(),
     Tags: [
       '#인천',
       '#여행',
@@ -35,12 +31,13 @@ export const photoCardInfo: PhotoCardInfoType[] = [
     ],
   },
   {
+    id: id++,
     ImageUrl: 'favicon.ico',
     Area: '인천 구월동',
     Like: 12,
     Text: '출력 잘되면 좋겠다',
     Writer: '글쓴이닉네임',
-    Date: '2022-04-08',
+    Date: new Date(),
     Tags: [
       '#인천',
       '#여행',
@@ -48,12 +45,13 @@ export const photoCardInfo: PhotoCardInfoType[] = [
     ],
   },
   {
+    id: id++,
     ImageUrl: 'favicon.ico',
     Area: '인천 구월동',
     Like: 12,
-    Text: '이건 뭔사진인지 당최 모르겠네 가나다라마바사아자차카타파하 아직도 남았는데 가나다라마바사 두줄이상이면 점으로 돼야하는데 이게 될라나',
+    Text: '출력 잘되면 좋겠다',
     Writer: '글쓴이닉네임',
-    Date: '2022-04-08',
+    Date: new Date(),
     Tags: [
       '#인천',
       '#여행',
