+++ src/utils/formatDate.ts
@@ -1,0 +1,4 @@
+export const formatDate = (date: Date) => {
+  const formattedDate = [ date.getFullYear(), date.getMonth() + 1, date.getDate()].join('-');
+  return formattedDate;
+};
